Validation

+++ SalesWebMvc/Controllers/SellersController.cs
@@ -37,6 +37,12 @@ namespace SalesWebMvc.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(Seller seller)
         {
+            var departments = _departmentService.FindAll();
+            var viewModel = new SellerFormViewModel { Departments = departments };
+            if (!ModelState.IsValid)
+            {
+                return View(viewModel);
+            }
             _sellerService.Insert(seller);
             return RedirectToAction(nameof(Index));
         }
@@ -59,6 +65,12 @@ namespace SalesWebMvc.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Delete(Seller seller)
         {
+            var departments = _departmentService.FindAll();
+            var viewModel = new SellerFormViewModel { Departments = departments };
+            if (!ModelState.IsValid)
+            {
+                return View(viewModel);
+            }
             _sellerService.Remove(seller.Id);
             return RedirectToAction(nameof(Index));
         }
